services/web/requirements.txt
Flask==1.1.2
user_agent
bs4
werkzeug==0.16.1
line-bot-sdk
datetime 
lxml
cchardet